Refactoring: - Added slug to leagues; - Changed seasons model to include the years it starts and ends; - Fixed fillable property on SeasonStanding model to properly save it's season_id; - Implemented a method to find leagues by slug; - Moved routes from api to web; - Created mixins to show toasts and confirm modals; - Saved the app url in a global variable in a script tag, it's no longer compiled with the code; New features: - Implemented a method to save seasons and seasonStandings; - Added vue-select to make multiple selections; - Added page titles to each page; - Seasons pages (list, form and inner);

+++ resources/js/components/TeamsTable.vue
@@ -21,8 +21,12 @@
 </template>
 
 <script>
+    import toasts from '../mixins/toasts';
+    import confirm from '../mixins/confirm';
+
     export default {
         name: "TeamsTable",
+        mixins: [toasts, confirm],
         props: {
             teams: {
                 type: Array,
@@ -41,12 +45,13 @@
                         key: 'id',
                         sortable: true,
                         label: '#',
-                        class: "id-column"
+                        class: "id-column align-row-center"
                     },
                     {
                         key: 'name',
                         sortable: true,
-                        label: 'Name'
+                        label: 'Name',
+                        class: 'align-row-center'
                     },
                     {
                         key: 'actions',
@@ -58,19 +63,17 @@
         },
         methods: {
             handleDelete(id) {
-                this.$bvModal.msgBoxConfirm(`Are you sure you want to delete team #${id}?`, {
-                    headerBgVariant: 'dark',
-                    footerBgVariant: 'dark',
-                    bodyBgVariant: 'dark',
+                const name = this.mtbTeams.find(i => i.id == id).name;
+
+                this.confirm({
+                    title: `Are you sure you want to delete '${name}'?`,
                     okVariant: 'danger',
                     okTitle: 'Delete'
-                })
-                    .then(value => {
-                        if (value) {
-                            this.deleteTeam(id);
-                        }
-                    })
-                    .catch();
+                }).then(value => {
+                    if (value) {
+                        this.deleteTeam(id);
+                    }
+                }).catch();
             },
             removeTeamFromTable(id) {
                 const index = this.mtbTeams.findIndex(i => i.id == id);
@@ -83,27 +86,15 @@
                     .delete('/teams/' + id)
                     .then(response => {
                         if (response.data.status) {
-                            this.$bvToast.toast('Team deleted successfully.', {
-                                title: 'Success!',
-                                variant: 'success',
-                                solid: true
-                            });
+                            this.addToastSuccess('Team deleted successfully.');
 
                             this.removeTeamFromTable(id);
                         }
                         else {
-                            this.$bvToast.toast('Error deleting team.', {
-                                title: 'Error!',
-                                variant: 'danger',
-                                solid: true
-                            });
+                            this.addToastError(null, 'Error deleting team.');
                         }
                     }).catch(error => {
-                        this.$bvToast.toast('Error deleting team.', {
-                            title: 'Error!',
-                            variant: 'danger',
-                            solid: true
-                        });
+                        this.addToastError(error, 'Error deleting team.');
                     });
             }
         },
